fix(probe): validate the property contract; drop the self-drawn window shadow
The _NET_WM_PID read trusted whatever type and format the property carried:
a malformed or hostile client can publish a format-8 property under the same
name, and the unconditional 8-byte read was both an over-wide read and a PID
mismatch risk. The read now demands CARDINAL/type with format 32 (host long
on LP64), reads exactly one element via ReadIntPtr, and treats everything
else as absent. The xvfb smoke's own XChangeProperty had the mirror-image
bug — it wrote int[] where format 32 requires host long[] — so the test that
proves the ABI was itself violating it.

The NativeAOT foundation test run now executes under xvfb-run too: the X11
probe path (and its ABI) is the product's shipping target, and it was only
ever exercised on CoreCLR.

The shell window drops its self-drawn shadow entirely. Per-pixel
transparency is not guaranteed on every Windows machine; when it fails, the
shadow paints over an opaque margin and exposes the rectangular window
bounds. The margin stays as a transparent buffer and XSR-703 now documents
the three size meanings (native outer, chrome surface, semantic viewport) so
acceptance tests name the one they measure.

+++ tests/PCL.Services.Tests/LaunchProgressTests.cs
@@ -775,7 +775,7 @@ internal static partial class Program
             nint white = X11TestApi.XWhitePixel(display, 0);
             nuint window = (nuint)X11TestApi.XCreateSimpleWindow(display, root, 0, 0, 64, 64, 0, black, white);
             nuint pidAtom = (nuint)X11TestApi.XInternAtom(display, "_NET_WM_PID", false);
-            int[] ownPid = [Environment.ProcessId];
+            nint[] ownPid = [Environment.ProcessId];
             _ = X11TestApi.XChangeProperty(
                 display, window, pidAtom, X11TestApi.XaCardinal, 32, X11TestApi.PropModeReplace, ownPid, 1);
